X00TD: add 3 more frequency steps to big and 1 to little cluster
This is highly experimental for now and might cause performance issues
since the big cluster is clocked at a lower frequency. Adding 300 Mhz
to the little cluster does not seem to make a performance drop in normal
usage. This also needs to be tested thoroughly. Voltage level
might need some adjusting.

Thanks to @uvera